allow to disable security - barcelona backend now works
A backend can in fact not have security at all -- sending any credential will
then fail, as our client side expects credentials to be honored.  The adaptor
now allows to continue without a valid context (more precise, with empty
contexts and with no contexts), and will in that case not include any credential
details in the request.